将自增ID修改为雪花ID❄️

+++ mdsp-common/src/main/java/com/yhcoo/common/constants/CommonConstants.java
@@ -6,7 +6,7 @@ public interface CommonConstants {
     /**
      * 树的根节点值
      */
-    Integer TREE_ROOT = -1;
+    Long TREE_ROOT = -1l;
 
     String MDSP_REDIS_LIST_LEY = "mdsp_client_id_to_access:cloud";
